# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Suomi API",
    "version": "1.0.0",
    "description": "Financial management API for couples"
  },
  "servers": [
    {
      "url": "/v1",
      "description": "API v1"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "paths": {
...
    "/v1/budget-categories": {
      "get": {
        "summary": "List all budget categories",
        "tags": [
          "Budget Categories"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of categories",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string"
                      },
                      "name": {
                        "type": "string"
                      },
                      "icon": {
                        "type": [
                          "string",
                          "null"
                        ]
                      },
                      "isSystem": {
                        "type": "boolean"
                      },
                      "createdAt": {
                        "type": "string"
                      }
                    },
                    "required": [
                      "id",
                      "name",
                      "icon",
                      "isSystem",
                      "createdAt"
                    ]
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid authentication token"
          }
        }
      },
      "post": {
        "summary": "Create a budget category",
        "tags": [
          "Budget Categories"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 100
                  },
                  "icon": {
                    "type": [
                      "string",
                      "null"
                    ]
                  }
                },
                "required": [
                  "name"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Category created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "name": {
                      "type": "string"
                    },
                    "icon": {
                      "type": [
                        "string",
                        "null"
                      ]
                    },
                    "isSystem": {
                      "type": "boolean"
                    },
                    "createdAt": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "id",
                    "name",
                    "icon",
                    "isSystem",
                    "createdAt"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Missing or invalid authentication token"
          }
        }
      }
    },
    "/v1/budget-categories/{id}": {
      "put": {
        "summary": "Update a budget category",
        "tags": [
          "Budget Categories"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 100
                  },
                  "icon": {
                    "type": [
                      "string",
                      "null"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Category updated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "name": {
                      "type": "string"
                    },
                    "icon": {
                      "type": [
                        "string",
                        "null"
                      ]
                    },
                    "isSystem": {
                      "type": "boolean"
                    },
                    "createdAt": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "id",
                    "name",
                    "icon",
                    "isSystem",
                    "createdAt"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid authentication token"
          },
          "404": {
            "description": "Category not found"
          }
        }
      },
      "delete": {
        "summary": "Delete a budget category",
        "tags": [
          "Budget Categories"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "204": {
            "description": "Category deleted"
          },
          "401": {
            "description": "Missing or invalid authentication token"
          },
          "404": {
            "description": "Category not found"
          }
        }
      }
    },
    "/v1/budget-plans": {
      "get": {
        "summary": "Get personal budget plan for a month",
        "tags": [
          "Budget Plans"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Personal plan with items",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "plan": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "customerId": {
                          "type": [
                            "string",
                            "null"
                          ]
                        },
                        "partnershipId": {
                          "type": [
                            "string",
                            "null"
                          ]
                        },
                        "yearMonth": {
                          "type": "string"
                        },
                        "currencyCode": {
                          "type": "string"
                        },
                        "isJoint": {
                          "type": "boolean"
                        },
                        "createdAt": {
                          "type": "string"
                        },
                        "updatedAt": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "id",
                        "customerId",
                        "partnershipId",
                        "yearMonth",
                        "currencyCode",
                        "isJoint",
                        "createdAt",
                        "updatedAt"
                      ]
                    },
                    "items": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "planId": {
                            "type": "string"
                          },
                          "categoryId": {
                            "type": "string"
                          },
                          "name": {
                            "type": "string"
                          },
                          "plannedAmount": {
                            "type": "integer"
                          },
                          "direction": {
                            "type": "string",
                            "enum": [
                              "INCOME",
                              "EXPENSE"
                            ]
                          },
                          "type": {
                            "type": "string",
                            "enum": [
                              "FIXED",
                              "ESTIMATED"
                            ]
                          },
                          "recurrence": {
                            "type": "string",
                            "enum": [
                              "PERMANENT",
                              "ONE_TIME",
                              "INSTALLMENT"
                            ]
                          },
                          "installmentTotal": {
                            "type": [
                              "number",
                              "null"
                            ]
                          },
                          "installmentNumber": {
                            "type": [
                              "number",
                              "null"
                            ]
                          },
                          "sourceItemId": {
                            "type": [
                              "string",
                              "null"
                            ]
                          },
                          "createdAt": {
                            "type": "string"
                          },
                          "updatedAt": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "id",
                          "planId",
                          "categoryId",
                          "name",
                          "plannedAmount",
                          "direction",
                          "type",
                          "recurrence",
                          "installmentTotal",
                          "installmentNumber",
                          "sourceItemId",
                          "createdAt",
                          "updatedAt"
                        ]
                      }
                    }
                  },
                  "required": [
                    "plan",
                    "items"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid authentication token"
          }
        }
      },
      "post": {
        "summary": "Create a personal budget plan",
        "tags": [
          "Budget Plans"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "yearMonth": {
                    "type": "string",
                    "pattern": "^\\d{4}-\\d{2}$"
                  },
                  "currencyCode": {
                    "type": "string",
                    "minLength": 3,
                    "maxLength": 3
                  }
                },
                "required": [
                  "yearMonth",
                  "currencyCode"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Plan created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "plan": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "customerId": {
                          "type": [
                            "string",
                            "null"
                          ]
                        },
                        "partnershipId": {
                          "type": [
                            "string",
                            "null"
                          ]
                        },
                        "yearMonth": {
                          "type": "string"
                        },
                        "currencyCode": {
                          "type": "string"
                        },
                        "isJoint": {
                          "type": "boolean"
                        },
                        "createdAt": {
                          "type": "string"
                        },
                        "updatedAt": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "id",
                        "customerId",
                        "partnershipId",
                        "yearMonth",
                        "currencyCode",
                        "isJoint",
                        "createdAt",
                        "updatedAt"
                      ]
                    },
                    "items": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "planId": {
                            "type": "string"
                          },
                          "categoryId": {
                            "type": "string"
                          },
                          "name": {
                            "type": "string"
                          },
                          "plannedAmount": {
                            "type": "integer"
                          },
                          "direction": {
                            "type": "string",
                            "enum": [
                              "INCOME",
                              "EXPENSE"
                            ]
                          },
                          "type": {
                            "type": "string",
                            "enum": [
                              "FIXED",
                              "ESTIMATED"
                            ]
                          },
                          "recurrence": {
                            "type": "string",
                            "enum": [
                              "PERMANENT",
                              "ONE_TIME",
                              "INSTALLMENT"
                            ]
                          },
                          "installmentTotal": {
                            "type": [
                              "number",
                              "null"
                            ]
                          },
                          "installmentNumber": {
                            "type": [
                              "number",
                              "null"
                            ]
                          },
                          "sourceItemId": {
                            "type": [
                              "string",
                              "null"
                            ]
                          },
                          "createdAt": {
                            "type": "string"
                          },
                          "updatedAt": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "id",
                          "planId",
                          "categoryId",
                          "name",
                          "plannedAmount",
                          "direction",
                          "type",
                          "recurrence",
                          "installmentTotal",
                          "installmentNumber",
                          "sourceItemId",
                          "createdAt",
                          "updatedAt"
                        ]
                      }
                    }
                  },
                  "required": [
                    "plan",
                    "items"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Missing or invalid authentication token"
          }
        }
      }
    },
    "/v1/budget-plans/{id}": {
      "delete": {
        "summary": "Delete a personal budget plan",
        "tags": [
          "Budget Plans"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "204": {
            "description": "Plan deleted"
          },
          "401": {
            "description": "Missing or invalid authentication token"
          },
          "404": {
            "description": "Plan not found"
          }
        }
      }
    },
...
    "/v1/partnership/budget-plans/{id}": {
      "delete": {
        "summary": "Delete a joint budget plan",
        "tags": [
          "Budget Plans"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "204": {
            "description": "Joint plan deleted"
          },
          "401": {
            "description": "Missing or invalid authentication token"
          },
          "404": {
            "description": "Plan not found"
          }
        }
      }
    },
    "/v1/budget-plans/summary": {
      "get": {
        "summary": "Get budget summary comparing planned vs actual",
        "tags": [
          "Budget Plans"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Budget summary"
          },
          "401": {
            "description": "Missing or invalid authentication token"
          }
        }
      }
    },
...
  },
...
}